Implemented a secure authentication layer using python-jose.

Key Changes:
Token Management: Users receive a JWT token with a 30-minute expiration upon successful login.

Route Protection: Added a gatekeeper dependency to all sensitive endpoints to ensure requests are authenticated.

Authorization: Enforced strict ownership logic so users can only perform CRUD actions on their own appointments.

Password Security: Verified login credentials against Argon2 hashes stored in the database.

This completes the security requirements for the v1 backend. 

Next steps: Connect the backend to flutter frontend and verify functionality

+++ backend/app/routes/appointment_router.py
@@ -2,8 +2,9 @@
 Router for HTTP requests related to appointment CRUD operations.
 """
 from datetime import datetime
+from typing import Annotated
 
-from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException, status, Response
 from sqlmodel import Session, select, extract
 
 from app.db import get_session
@@ -16,37 +17,29 @@ from app.models.appointment import (
 )
 from app.services.distance_service import get_or_create_distance
 from app.services.pdf_service import generate_pdf_report
+from app.core.security import get_current_user
 
-appointment_router = APIRouter(tags=["appointments"])
+
+appointment_router = APIRouter(prefix='/appointments', tags=["appointments"], )
 
 
-@appointment_router.post(
-    "/users/{user_id}/appointments",
-    response_model=AppointmentPublic,
-    status_code=status.HTTP_201_CREATED,
-)
+@appointment_router.post("/", response_model=AppointmentPublic)
 def create_appointment(
-    user_id: int,
     appointment_data: AppointmentCreate,
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
 ):
-    user = session.get(User, user_id)
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
-
-
+    """
+    Creates an appointment by taking in an AppointmentCreate object 
+    """
     distance = get_or_create_distance(
         session=session,
-        origin_address=user.address,
+        origin_address=current_user.address,
         destination_address=appointment_data.destination_address,
     )
 
     extra_data = {
-        'user_id': user.id,
+        'user_id': current_user.id,
         'distance_id': distance.id,
         'roundtrip_distance': distance.roundtrip_distance
     }
@@ -60,11 +53,26 @@ def create_appointment(
     return new_appointment
 
 
-@appointment_router.get("/appointments/{appointment_id}", response_model=AppointmentPublic)
-def get_appointment(
-    appointment_id: int,
+@appointment_router.get('/', response_model=list[AppointmentPublic])
+def get_appointment_list(
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
 ):
+    appointments = session.exec(
+        select(Appointment)
+        .where(Appointment.user_id == current_user.id)
+        .order_by(Appointment.appointment_date)
+    ).all()
+
+    return appointments
+
+
+@appointment_router.get("/{appointment_id}", response_model=AppointmentPublic)
+def get_appointment(
+    appointment_id: int,
+    current_user: Annotated[User, Depends(get_current_user)],
+    session: Session = Depends(get_session),
+):    
     appointment = session.get(Appointment, appointment_id)
 
     if not appointment:
@@ -72,43 +80,44 @@ def get_appointment(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Appointment not found",
         )
+    
+    if appointment.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail='Unauthorized access'
+            )
 
     return appointment
 
 
-@appointment_router.get("/users/{user_id}/appointments/reports")
+@appointment_router.get("/reports")
 def get_user_appointments_report(
-    user_id: int,
-    year: int | None = datetime.now().year,
+    current_user: Annotated[User, Depends(get_current_user)],
+    year: int | None = None,
     session: Session = Depends(get_session),
-):
-    user = session.get(User, user_id)
-
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+) -> Response:
+    report_year = year or datetime.now().year
 
     appointments = session.exec(
         select(Appointment)
-        .where(Appointment.user_id == user_id)
-        .where(extract('year', Appointment.appointment_date) == year)
+        .where(Appointment.user_id == current_user.id)
+        .where(extract('year', Appointment.appointment_date) == report_year)
         .order_by(Appointment.appointment_date, Appointment.id)
     ).all()
 
-    if len(appointments) == 0:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'No appointments found for {year}')
+    if not appointments:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'No appointments found for {report_year}')
     
-    return generate_pdf_report(appointments, user, year)
+    return generate_pdf_report(appointments, current_user, report_year)
 
 
-@appointment_router.patch("/appointments/{appointment_id}", response_model=AppointmentPublic)
+@appointment_router.patch("/{appointment_id}", response_model=AppointmentPublic)
 def update_appointment(
     appointment_id: int,
     appointment_data: AppointmentUpdate,
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
-):
+):   
     appointment = session.get(Appointment, appointment_id)
 
     if not appointment:
@@ -116,26 +125,23 @@ def update_appointment(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Appointment not found",
         )
+    
+    if appointment.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail='Unauthorized access'
+            )
 
     appointment_data_dump = appointment_data.model_dump(exclude_unset=True)
 
-    if "destination_address" in appointment_data_dump:
-        user = session.get(User, appointment.user_id)
+    distance = get_or_create_distance(
+        session=session,
+        origin_address=current_user.address,
+        destination_address=appointment_data_dump["destination_address"],
+    )
 
-        if not user:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Associated user not found",
-            )
-
-        distance = get_or_create_distance(
-            session=session,
-            origin_address=user.address,
-            destination_address=appointment_data_dump["destination_address"],
-        )
-
-        appointment.distance_id = distance.id
-        appointment.roundtrip_distance = distance.roundtrip_distance
+    appointment.distance_id = distance.id
+    appointment.roundtrip_distance = distance.roundtrip_distance
 
     appointment.sqlmodel_update(appointment_data_dump)
 
@@ -146,23 +152,21 @@ def update_appointment(
     return appointment
 
 
-@appointment_router.delete("/appointments/{appointment_id}")
+@appointment_router.delete("/{appointment_id}")
 def delete_appointment(
     appointment_id: int,
+    current_user: Annotated[User, Depends(get_current_user)],
     session: Session = Depends(get_session),
 ):
     appointment = session.get(Appointment, appointment_id)
 
     if not appointment:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Appointment not found",
-        )
+        raise HTTPException(status_code=404, detail="Appointment not found")
+    
+    if appointment.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail='Access forbidden')
 
     session.delete(appointment)
     session.commit()
 
-    return {
-        "message": "Appointment deleted",
-        "appointment_id": appointment_id,
-    }
+    return {"message": "Appointment deleted", "appointment_id": appointment_id}
